+++ test/Peast/Syntax/ES2015/files/Misc/ContinueBreakNoLabelNewLineAfter.js
@@ -1,0 +1,6 @@
+for (;;) {
+    if (true) break
+        ; else {}
+    if (true) continue
+        ; else {}
+}